I was thinking, would it be best to work on the roles & permissions part of the 'profile'?
For example, if a marketing person wants to be a 'creator' (creates, reviews, A/B views and saves landing pages) should we provide this in the 'Signup' process? I was thinking we'd add a couple of questions in the signup process so we can determine the authentication level for each new user?